requirements.txt
selenium
attrs
flake8
flake8-colors
pylint
pytest
pytest-cov
coveralls
